Fix many compiler warnings, and minor code cleanup

+++ source/xen/sren/renderer3d_rasterizer.cpp
@@ -159,7 +159,7 @@ namespace xen{
 
 			// Find the actual view_region we wish to draw to. This is the
 			// intersection of the actual target, and the user specified viewport
-			xen::Aabb2u screen_rect = { 0, 0, target.width - 1, target.height - 1 };
+			xen::Aabb2u screen_rect = { 0, 0, (u32)target.width - 1, (u32)target.height - 1 };
 			xen::Aabb2r view_region = (xen::Aabb2r)xen::getIntersection(viewport, screen_rect);
 
 			Mat4r mat_vp = xen::getViewProjectionMatrix(params.camera, view_region.max - view_region.min);
